requirements.txt
fastapi
uvicorn[standard]
python-multipart
chardet
pdfminer.six
python-docx
striprtf